+++ demos/grass-skinning-gpu-instancing/PaleGL/shaders/skinningShader.js
@@ -1,11 +1,3 @@
-// import {transformVertexUniforms, shadowMapVertexUniforms} from "./commonUniforms";
-// import {shadowMapVertex} from "./shadowMapShader";
-
-// export const skinningVertexAttributes = (beginIndex) => [
-// `layout(location = ${beginIndex + 0}) in vec4 aBoneIndices;`,
-// `layout(location = ${beginIndex + 1}) in vec4 aBoneWeights;`,
-// ];
-
 export const calcSkinningMatrixFunc = () => `
 mat4 calcSkinningMatrix(mat4 jointMat0, mat4 jointMat1, mat4 jointMat2, mat4 jointMat3, vec4 boneWeights) {
     mat4 skinMatrix =
@@ -22,7 +14,7 @@ struct SkinAnimationClipData {
     int frameCount;
 };
 
-mat4 getJointMatrix(sampler2D jointTexture, int jointIndex, int colNum) {
+mat4 getJointMatrix(sampler2D jointTexture, uint jointIndex, int colNum) {
     int colIndex = int(mod(float(jointIndex), float(colNum))); // 横
     int rowIndex = int(floor(float(jointIndex) / float(colNum))); // 縦
     mat4 jointMatrix = mat4(
@@ -42,7 +34,7 @@ mat4 getJointMatrix(sampler2D jointTexture, int jointIndex, int colNum) {
 
 mat4 getJointMatrixGPUSkinning(
     sampler2D jointTexture,
-    int jointIndex,
+    uint jointIndex,
     int jointNum,
     int currentSkinIndex,
     int colNum,
@@ -50,9 +42,9 @@ mat4 getJointMatrixGPUSkinning(
     float time,
     float timeOffset
 ) {
-    int offset = int(mod(floor(time + timeOffset), float(totalFrameCount))) * jointNum;
-    int colIndex = int(mod(float(jointIndex + offset), float(colNum))); // 横
-    int rowIndex = int(floor(float(jointIndex + offset) / float(colNum))); // 縦
+    float offset = float(int(mod(floor(time + timeOffset), float(totalFrameCount))) * jointNum);
+    int colIndex = int(mod(float(jointIndex) + offset, float(colNum))); // 横
+    int rowIndex = int(floor(float(jointIndex) + offset / float(colNum))); // 縦
 
     mat4 jointMatrix = mat4(
         texelFetch(jointTexture, ivec2(colIndex * 4 + 0, rowIndex), 0),
@@ -87,10 +79,10 @@ export const skinningVertex = (gpuSkinning = false) => `
     ${gpuSkinning ? `
     // gpu skinning
     float fps = 30.;
-    mat4 jointMatrix0 = getJointMatrixGPUSkinning(uJointTexture, int(aBoneIndices[0]), uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
-    mat4 jointMatrix1 = getJointMatrixGPUSkinning(uJointTexture, int(aBoneIndices[1]), uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
-    mat4 jointMatrix2 = getJointMatrixGPUSkinning(uJointTexture, int(aBoneIndices[2]), uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
-    mat4 jointMatrix3 = getJointMatrixGPUSkinning(uJointTexture, int(aBoneIndices[3]), uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
+    mat4 jointMatrix0 = getJointMatrixGPUSkinning(uJointTexture, aBoneIndices[0], uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
+    mat4 jointMatrix1 = getJointMatrixGPUSkinning(uJointTexture, aBoneIndices[1], uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
+    mat4 jointMatrix2 = getJointMatrixGPUSkinning(uJointTexture, aBoneIndices[2], uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
+    mat4 jointMatrix3 = getJointMatrixGPUSkinning(uJointTexture, aBoneIndices[3], uBoneCount, 0, uJointTextureColNum, uTotalFrameCount, uTime * fps, aInstanceAnimationOffset);
     mat4 skinMatrix = calcSkinningMatrix(
         jointMatrix0,
         jointMatrix1,
@@ -99,10 +91,10 @@ export const skinningVertex = (gpuSkinning = false) => `
         aBoneWeights
     );
     ` : `
-    mat4 jointMatrix0 = getJointMatrix(uJointTexture, int(aBoneIndices[0]), uJointTextureColNum);
-    mat4 jointMatrix1 = getJointMatrix(uJointTexture, int(aBoneIndices[1]), uJointTextureColNum);
-    mat4 jointMatrix2 = getJointMatrix(uJointTexture, int(aBoneIndices[2]), uJointTextureColNum);
-    mat4 jointMatrix3 = getJointMatrix(uJointTexture, int(aBoneIndices[3]), uJointTextureColNum);
+    mat4 jointMatrix0 = getJointMatrix(uJointTexture, aBoneIndices[0], uJointTextureColNum);
+    mat4 jointMatrix1 = getJointMatrix(uJointTexture, aBoneIndices[1], uJointTextureColNum);
+    mat4 jointMatrix2 = getJointMatrix(uJointTexture, aBoneIndices[2], uJointTextureColNum);
+    mat4 jointMatrix3 = getJointMatrix(uJointTexture, aBoneIndices[3], uJointTextureColNum);
     mat4 skinMatrix = calcSkinningMatrix(
         jointMatrix0,
         jointMatrix1,
